+++ react-forms/src/components/controlled-form/components/index.ts
@@ -1,0 +1,8 @@
+export { Age } from './age';
+export { Country } from './country';
+export { Email } from './email';
+export { Gender } from './gender';
+export { Name } from './name';
+export { Password } from './password';
+export { TermsAgreement } from './terms-agreement';
+export { UploadImage } from './upload-image';
