+++ react-forms/src/components/controlled-form/components/password.tsx
@@ -1,5 +1,5 @@
 import { ComponentsProps } from '../types';
-import styles from '../form.module.css';
+import styles from '../../../shared/form.module.css';
 
 export function Password({ register, errors }: ComponentsProps) {
   return (
