+++ react-forms/src/components/controlled-form/components/terms-agreement.tsx
@@ -1,5 +1,5 @@
 import { ComponentsProps } from '../types';
-import styles from '../form.module.css';
+import styles from '../../../shared/form.module.css';
 
 export function TermsAgreement({ register, errors }: ComponentsProps) {
   return (
