+++ react-forms/src/components/uncontrolled-form/components/index.ts
@@ -1,0 +1,2 @@
+export { Country } from './country';
+export { UploadImage } from './upload-image';
