+++ react-forms/src/shared/hooks/index.ts
@@ -1,0 +1,3 @@
+export { useClickOutside } from './useClickOutside';
+export { useEscapeKey } from './useEscapeKey';
+export { useModal } from './useModal';
