+++ react-forms/src/shared/store/index.ts
@@ -1,0 +1,7 @@
+export {
+  clearImage,
+  setInfo,
+  setImage,
+  useAppDispatch,
+  useAppSelector,
+} from './store';
