--- react-forms/src/types.ts
@@ -1,15 +1,0 @@
-export interface Data {
-  age: number;
-  confirmPassword: string;
-  country: string;
-  email: string;
-  firstName: string;
-  gender: string;
-  password: string;
-  terms: boolean;
-}
-
-export interface SetImagePayload {
-  data: string;
-  name: string;
-}
